15.0.0

+++ controllers/cards.js
@@ -1,9 +1,10 @@
 const Card = require('../models/card');
+const NotFoundError = require('../errors/not-found-err');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => { res.status(200).send(cards); })
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch(next);
 };
 
 const createCard = (req, res) => {
@@ -11,52 +12,59 @@ const createCard = (req, res) => {
   Card.create({ name, link, owner: req.user._id })
     .then((card) => {
       if (!card) {
-        return res.status(404).send({ message: 'Данные карточки не пришли' });
+        throw new NotFoundError('Данные карточки не пришли');
       }
 
       return res.send({ data: card });
     })
-    .catch(() => res.status(400).send({ message: 'Произошла ошибка при создании карточки' }));
+    .catch(next);
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.id)
+const deleteCard = (req, res, next) => {
+  Card.findById(req.params.id)
     .then((card) => {
       if (!card) {
-        return res.status(404).send({ message: 'Карточка не найдена' });
+        throw new NotFoundError('Карточка не найдена');
       }
-      return res.send({ data: card });
+      if(String(card.owner) !== req.user._id) {
+        throw new NotFoundError('Нельзя удалять чужие карточки');
+      }
+
+      Card.findByIdAndRemove(req.params.id)
+        .then(() => {
+          return res.send({ data: card });
+        })
     })
-    .catch(() => { res.status(400).send({ message: 'Произошла ошибка при удалении карточки' }); });
+    .catch(next);
 };
 
 // лайки
 
-const addLikeCard = (req, res) => Card.findByIdAndUpdate(
+const addLikeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
   { new: true },
 )
   .then((card) => {
     if (!card) {
-      return res.status(404).send({ message: 'Карточка не найдена' });
+      throw new NotFoundError('Карточка не найдена' );
     }
     return res.send({ data: card });
   })
-  .catch(() => { res.status(400).send({ message: 'Произошла ошибка при лайке карточки' }); });
+  .catch(next);
 
-const removeLikeCard = (req, res) => Card.findByIdAndUpdate(
+const removeLikeCard = (req, res, next) => Card.findByIdAndUpdate(
   req.params.cardId,
   { $pull: { likes: req.user._id } }, // убрать _id из массива
   { new: true },
 )
   .then((card) => {
     if (!card) {
-      return res.status(404).send({ message: 'Карточка не найдена' });
+      throw new NotFoundError('Карточка не найдена');
     }
     return res.send({ data: card });
   })
-  .catch(() => { res.status(404).send({ message: 'Произошла ошибка при убирании лайка' }); });
+  .catch(next);
 
 module.exports = {
   getCards, createCard, deleteCard, addLikeCard, removeLikeCard,
